requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
asyncpg==0.29.0
python-dotenv==1.0.1
